[FLN-42][Hoang][UI] Rename folder for ticket FLN-43 #26.05.2025
