+++ FrontEnd/src/assets/AdminHeaderAndSidebar/Sidebar.css
@@ -4,7 +4,7 @@
   color: #b0b8c1;
   display: flex;
   flex-direction: column;
-  height: 100vh;
+  /* height: 100; */
   min-height: unset;
   overflow: hidden;
 }
